requirements.txt
gitpython
docker
